MWPW-167050 Acrobat Batch 1b (#314)
This PR adds the following changes:
- Refactor Acrobat Workflow for scalability
- Config updates for each batch 1b verb
- Modularize the transition screen

Resolves: [MWPW-167050](https://jira.corp.adobe.com/browse/MWPW-167050)

**Test URLs:**
- Before:
-
https://stage--dc--adobecom.hlx.page/acrobat/online/test/delete-pdf-pages?unitylibs=stage
-
https://stage--dc--adobecom.hlx.page/acrobat/online/test/add-pages-to-pdf?unitylibs=stage
-
https://stage--dc--adobecom.hlx.page/acrobat/online/test/rearrange-pdf?unitylibs=stage
-
https://stage--dc--adobecom.hlx.page/acrobat/online/test/extract-pdf-pages?unitylibs=stage
-
https://stage--dc--adobecom.hlx.page/acrobat/online/test/request-signature?unitylibs=stage
- After:
-
https://stage--dc--adobecom.hlx.page/acrobat/online/test/delete-pdf-pages?unitylibs=batch1b
-
https://stage--dc--adobecom.hlx.page/acrobat/online/test/add-pages-to-pdf?unitylibs=batch1b
-
https://stage--dc--adobecom.hlx.page/acrobat/online/test/rearrange-pdf?unitylibs=batch1b
-
https://stage--dc--adobecom.hlx.page/acrobat/online/test/extract-pdf-pages?unitylibs=batch1b
-
https://stage--dc--adobecom.hlx.page/acrobat/online/test/request-signature?unitylibs=batch1b

---------
