+++ unitylibs/core/workflow/workflow-acrobat/target-config.json
@@ -1,285 +1,17 @@
 {
-  "verb-widget.fillsign": {
-    "type": "pdf",
+  "verb-widget": {
     "selector": ".verb-wrapper",
-    "handler": "render",
     "renderWidget": false,
     "source": ".verb-wrapper .verb-container",
     "target": ".verb-wrapper .verb-container",
-    "limits": {
-      "maxNumFiles": 1,
-      "maxFileSize": 104857600,
-      "pageLimit": {
-        "maxNumPages": 100
-      },
-      "allowedFileTypes": ["application/pdf"]
-    },
     "showSplashScreen": true,
     "splashScreenConfig": {
       "fragmentLink": "/dc-shared/fragments/shared-fragments/frictionless/splash-page/splashscreen",
       "splashScreenParent": "body"
     },
     "actionMap": {
-      ".verb-wrapper": [
-        {
-          "actionType": "fillsign"
-        },
-        {
-          "actionType": "continueInApp"
-        }
-      ],
-      "#file-upload": [
-        {
-          "actionType": "fillsign"
-        },
-        {
-          "actionType": "continueInApp"
-        }
-      ]
-    }
-  },
-  "verb-widget.compress-pdf": {
-    "type": "pdf",
-    "selector": ".verb-wrapper",
-    "handler": "render",
-    "renderWidget": false,
-    "source": ".verb-wrapper .verb-container",
-    "target": ".verb-wrapper .verb-container",
-    "limits": {
-      "maxFileSize": 2147483648,
-      "allowedFileTypes": [
-        "application/pdf",
-        "application/msword",
-        "application/xml",
-        "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-        "application/vnd.ms-powerpoint",
-        "application/vnd.openxmlformats-officedocument.presentationml.presentation",
-        "application/x-tika-ooxml",
-        "application/vnd.ms-excel",
-        "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-        "application/x-tika-msworks-spreadsheet",
-        "application/vnd.adobe.form.fillsign",
-        "application/illustrator",
-        "application/rtf",
-        "application/x-indesign",
-        "image/jpeg",
-        "image/png",
-        "image/bmp",
-        "image/gif",
-        "image/vnd.adobe.photoshop",
-        "image/tiff",
-        "message/rfc822",
-        "text/plain"
-      ]
-    },
-    "showSplashScreen": true,
-    "splashScreenConfig": {
-      "fragmentLink": "/dc-shared/fragments/shared-fragments/frictionless/splash-page/splashscreen",
-      "splashScreenParent": "body"
-    },
-    "actionMap": {
-      ".verb-wrapper": [
-        {
-          "actionType": "compress"
-        },
-        {
-          "actionType": "continueInApp"
-        }
-      ],
-      "#file-upload": [
-        {
-          "actionType": "compress"
-        },
-        {
-          "actionType": "continueInApp"
-        }
-      ]
-    }
-  },
-  "verb-widget.add-comment": {
-    "type": "pdf",
-    "selector": ".verb-wrapper",
-    "handler": "render",
-    "renderWidget": false,
-    "source": ".verb-wrapper .verb-container",
-    "target": ".verb-wrapper .verb-container",
-    "limits": {
-      "maxNumFiles": 1,
-      "maxFileSize": 104857600,
-      "allowedFileTypes": ["application/pdf"]
-    },
-    "showSplashScreen": true,
-    "splashScreenConfig": {
-      "fragmentLink": "/dc-shared/fragments/shared-fragments/frictionless/splash-page/splashscreen",
-      "splashScreenParent": "body"
-    },
-    "actionMap": {
-      ".verb-wrapper": [
-        {
-          "actionType": "addcomment"
-        },
-        {
-          "actionType": "continueInApp"
-        }
-      ],
-      "#file-upload": [
-        {
-          "actionType": "addcomment"
-        },
-        {
-          "actionType": "continueInApp"
-        }
-      ]
-    }
-  },
-  "verb-widget.split-pdf": {
-    "type": "pdf",
-    "selector": ".verb-wrapper",
-    "handler": "render",
-    "renderWidget": false,
-    "source": ".verb-wrapper .verb-container",
-    "target": ".verb-wrapper .verb-container",
-    "limits": {
-      "maxNumFiles": 1,
-      "maxFileSize": 1073741824,
-      "pageLimit": {
-        "minNumPages": 2,
-        "maxNumPages": 1500
-      },
-      "allowedFileTypes": ["application/pdf"],
-      "signedInallowedFileTypes": [
-        "application/msword",
-        "application/xml",
-        "application/vnd.openxmlformats-officedocument.wordprocessingml.document",
-        "application/vnd.ms-powerpoint",
-        "application/vnd.openxmlformats-officedocument.presentationml.presentation",
-        "application/x-tika-ooxml",
-        "application/vnd.ms-excel",
-        "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet",
-        "application/x-tika-msworks-spreadsheet",
-        "application/vnd.adobe.form.fillsign",
-        "application/illustrator",
-        "application/rtf",
-        "application/x-indesign",
-        "image/jpeg",
-        "image/png",
-        "image/bmp",
-        "image/gif",
-        "image/vnd.adobe.photoshop",
-        "image/tiff",
-        "message/rfc822",
-        "text/plain"
-      ]
-    },
-    "showSplashScreen": true,
-    "splashScreenConfig": {
-      "fragmentLink": "/dc-shared/fragments/shared-fragments/frictionless/splash-page/splashscreen",
-      "splashScreenParent": "body"
-    },
-    "actionMap": {
-      ".verb-wrapper": [
-        {
-          "actionType": "splitpdf"
-        },
-        {
-          "actionType": "continueInApp"
-        }
-      ],
-      "#file-upload": [
-        {
-          "actionType": "splitpdf"
-        },
-        {
-          "actionType": "continueInApp"
-        }
-      ]
-    }
-  },
-  "verb-widget.number-pages": {
-    "type": "pdf",
-    "selector": ".verb-wrapper",
-    "handler": "render",
-    "renderWidget": false,
-    "source": ".verb-wrapper .verb-container",
-    "target": ".verb-wrapper .verb-container",
-    "limits": {
-      "maxNumFiles": 1,
-      "maxFileSize": 104857600,
-      "allowedFileTypes": ["application/pdf"]
-    },
-    "showSplashScreen": true,
-    "splashScreenConfig": {
-      "fragmentLink": "/dc-shared/fragments/shared-fragments/frictionless/splash-page/splashscreen",
-      "splashScreenParent": "body"
-    },
-    "actionMap": {
-      ".verb-wrapper": [
-        {
-          "actionType": "numberpages"
-        },
-        {
-          "actionType": "continueInApp"
-        }
-      ],
-      "#file-upload": [
-        {
-          "actionType": "numberpages"
-        },
-        {
-          "actionType": "continueInApp"
-        }
-      ]
-    }
-  },
-  "verb-widget.crop-pages": {
-    "type": "pdf",
-    "selector": ".verb-wrapper",
-    "handler": "render",
-    "renderWidget": false,
-    "source": ".verb-wrapper .verb-container",
-    "target": ".verb-wrapper .verb-container",
-    "limits": {
-      "maxNumFiles": 1,
-      "maxFileSize": 104857600,
-      "allowedFileTypes": ["application/pdf"]
-    },
-    "showSplashScreen": true,
-    "splashScreenConfig": {
-      "fragmentLink": "/dc-shared/fragments/shared-fragments/frictionless/splash-page/splashscreen",
-      "splashScreenParent": "body"
-    },
-    "actionMap": {
-      ".verb-wrapper": [
-        {
-          "actionType": "croppages"
-        },
-        {
-          "actionType": "continueInApp"
-        }
-      ],
-      "#file-upload": [
-        {
-          "actionType": "croppages"
-        },
-        {
-          "actionType": "continueInApp"
-        }
-      ]
-    }
-  },
-  "marquee": {
-    "type": "pdf",
-    "selector": ".action-area",
-    "handler": "render",
-    "renderWidget": false,
-    "source": ".action-area .con-button",
-    "target": ".action-area .con-button",
-    "actionMap": {
-      ".action-area .con-button": [
-        {
-          "actionType": "upload"
-        }
-      ]
+      ".verb-wrapper": "upload",
+      "#file-upload": "upload"
     }
   }
 }
