+++ unitylibs/core/workflow/workflow.js
@@ -176,7 +176,6 @@ class WfInitiator {
     } else {
       this.actionMap = this.targetConfig.actionMap;
     }
-    this.limits = this.targetConfig.limits;
     const { default: ActionBinder } = await import(`${getUnityLibs()}/core/workflow/${this.workflowCfg.name}/action-binder.js`);
     await new ActionBinder(
       this.el,
@@ -184,7 +183,6 @@ class WfInitiator {
       this.targetBlock,
       this.interactiveArea,
       this.actionMap,
-      this.limits,
     ).initActionListeners();
   }
 
@@ -273,6 +271,11 @@ class WfInitiator {
           'number-pages',
           'split-pdf',
           'crop-pages',
+          'delete-pages',
+          'insert-pdf',
+          'extract-pages',
+          'reorder-pages',
+          'sendforsignature'
         ]),
       },
       'workflow-ai': {
